pyproject.toml
[project]
name = "semantic-drive-search"
version = "0.2.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-genai>=1.0.0",
    "google-api-python-client>=2.160.0",
    "google-auth-oauthlib>=1.2.0",
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "itsdangerous>=2.2.0",
    "starlette[full]>=0.45.0",
    "httpx>=0.28.0",
    "typer>=0.12.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
]

[project.scripts]
sds = "backend.cli:app"
sds-mcp = "backend.mcp_server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.28.0",
    "ruff>=0.4.0",
    "black>=24.4.0",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "ARG001", # unused function argument
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG", "S101"]

# Black configuration
[tool.black]
target-version = ["py312"]
line-length = 100

# MyPy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["backend"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
